[Finishes #163978053] Restrecturing the folder structure
