+++ server/__init__.py
@@ -1,7 +1,6 @@
 """ Server initialization module """
 import os
 from flask import Flask
-from server.api.v1.admin.admin_views import admin_api_v1_blueprint
 
 app = Flask(__name__)
 
@@ -12,4 +11,8 @@ SERVER_SETTINGS = os.getenv(
 
 app.config.from_object('server.config.DevelopmentConfig')
 
-app.register_blueprint(admin_api_v1_blueprint, url_prefix='/api/v1')
+from server.api.v1.views.office import office_api_v1_blueprint
+from server.api.v1.views.party import party_api_v1_blueprint
+
+app.register_blueprint(office_api_v1_blueprint, url_prefix='/api/v1')
+app.register_blueprint(party_api_v1_blueprint, url_prefix='/api/v1')
