+++ server/api/tests/v1/test_parties.py
