+++ server/api/v1/models/office.py
@@ -1,0 +1,15 @@
+""" Office Models """
+offices_data = []
+ 
+
+class Offices():
+    """ Class for Offices data """
+    def __init__(self):
+        
+        self.offices_data = offices_data
+
+
+    def save(self):
+        """ Save Office method """
+        offices_data.append(self)
+
