+++ server/api/v1/models/party.py
@@ -1,0 +1,12 @@
+""" Office Models """
+parties_data = []
+
+
+class Parties():
+    """ Class for parties data """
+    def __init__(self):
+        self.parties_data = parties_data
+
+    def save(self):
+        """ Save party method"""
+        parties_data.append(self)
